+++ config/enso/config.php
@@ -1,7 +1,7 @@
 <?php
 
 return [
-    'version' => '2.6.8',
+    'version' => '2.6.9',
     'defaultRole' => 'admin',
     'phpDateFormat' => 'd-m-Y',
     'jsDateFormat' => 'DD-MM-YYYY',