+++ resources/assets/js/components/enso/vuedatatable/TopControls.vue
@@ -146,6 +146,11 @@ export default {
 
             if (button.action === 'router') {
                 this.$router.push({ name: button.route });
+                return;
+            }
+
+            if (button.action === 'ajax') {
+                this.$emit('action', button.method, button.path, button.postEvent);
             }
         },
     },